Use Mailbox call to determine processor types
We were using the bcm_host library for this staff, which refers to the device
tree to obtain processor-specific identifier.  However, the information is
silently appended to the device tree blob by the firmware, and some bootloaders
including U-Boot tend to fail to handle that.  So we pick out a more robust way.